pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "preprompt"
version = "0.1.0"
description = "MCP server that intercepts and optimizes prompts in Claude Code and Cursor"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["mcp", "llm", "prompt-engineering", "claude", "cursor", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "mcp>=1.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/yashdeeptehlan/preprompt"
Repository = "https://github.com/yashdeeptehlan/preprompt"
Issues = "https://github.com/yashdeeptehlan/preprompt/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
dist = [
    "build",
    "twine",
]

[project.scripts]
preprompt                    = "mcp_server.server:main"
preprompt-history            = "cli.commands:history_cmd"
preprompt-stats              = "cli.commands:stats_cmd"
preprompt-test-classifier    = "cli.commands:test_classifier_cmd"
preprompt-memory             = "cli.commands:memory_cmd"
preprompt-update-context     = "cli.commands:update_context_cmd"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "storage", "cli"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
